+++ proj-java/src/main/java/common/Question.java
@@ -18,7 +18,7 @@ public class Question extends Post{
 	* Construtor parametrizado
 	*/
 	public Question(String title, int nanswers, Map<Integer, String> tags, LocalDate lastActivityDate,
-									char postTypeId, int id, int ownerUserId, LocalDate creationDate){
+									char postTypeId, long id, long ownerUserId, LocalDate creationDate){
 		super(postTypeId, id, ownerUserId, creationDate);
 		this.title = title;
 		this.nanswers = nanswers;
